Вынес стартовую команду отдельно + статик импорты для анотаций
