+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/BSLLSPLauncher.java
@@ -21,104 +21,14 @@
  */
 package com.github._1c_syntax.bsl.languageserver;
 
-import com.github._1c_syntax.bsl.languageserver.cli.AnalyzeCommand;
-import com.github._1c_syntax.bsl.languageserver.cli.FormatCommand;
-import com.github._1c_syntax.bsl.languageserver.cli.VersionCommand;
-import com.github._1c_syntax.bsl.languageserver.configuration.LanguageServerConfiguration;
-import lombok.extern.slf4j.Slf4j;
-import org.eclipse.lsp4j.jsonrpc.Launcher;
-import org.eclipse.lsp4j.launch.LSPLauncher;
-import org.eclipse.lsp4j.services.LanguageClient;
-import org.eclipse.lsp4j.services.LanguageClientAware;
-import org.eclipse.lsp4j.services.LanguageServer;
+import com.github._1c_syntax.bsl.languageserver.cli.LanguageServerStartCommand;
 import picocli.CommandLine;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.PrintWriter;
-import java.io.UnsupportedEncodingException;
-import java.nio.charset.StandardCharsets;
-import java.util.concurrent.Callable;
-
-@Slf4j
-@CommandLine.Command(
-  name = "",
-  description = "BSL language server on LSP server mode",
-  subcommands = {
-    AnalyzeCommand.class,
-    FormatCommand.class,
-    VersionCommand.class
-  },
-  usageHelpAutoWidth = true,
-  synopsisSubcommandLabel = "[COMMAND [ARGS]]",
-  footer = "@|green Copyright(c) 2018-2020|@",
-  header = "@|green BSL language server|@")
-public class BSLLSPLauncher implements Callable<Integer> {
-
-  @CommandLine.Option(
-    names = {"-h", "--help"},
-    usageHelp = true,
-    description = "Show this help message and exit")
-  boolean usageHelpRequested;
-
-  @CommandLine.Option(
-    names = {"-c", "--configuration"},
-    description = "Path to language server configuration file",
-    paramLabel = "<path>",
-    defaultValue = "")
-  private String configurationOption;
-
+public class BSLLSPLauncher {
   public static void main(String[] args) {
-
-    int result = new CommandLine(new BSLLSPLauncher()).execute(args);
+    int result = new CommandLine(new LanguageServerStartCommand()).execute(args);
     if (result >= 0) {
       System.exit(result);
     }
-  }
-
-  public Integer call() {
-    File configurationFile = new File(configurationOption);
-
-    LanguageServerConfiguration configuration = LanguageServerConfiguration.create(configurationFile);
-    LanguageServer server = new BSLLanguageServer(configuration);
-
-    Launcher<LanguageClient> launcher = getLanguageClientLauncher(server, configuration);
-
-    LanguageClient client = launcher.getRemoteProxy();
-    ((LanguageClientAware) server).connect(client);
-
-    launcher.startListening();
-    return -1;
-  }
-
-  private static Launcher<LanguageClient> getLanguageClientLauncher(
-    LanguageServer server,
-    LanguageServerConfiguration configuration
-  ) {
-    InputStream in = System.in;
-    OutputStream out = System.out;
-
-    File logFile = configuration.getTraceLog();
-    if (logFile == null) {
-      return LSPLauncher.createServerLauncher(server, in, out);
-    }
-
-    Launcher<LanguageClient> launcher;
-
-    try {
-      PrintWriter printWriter = new PrintWriter(logFile, StandardCharsets.UTF_8.name());
-      launcher = LSPLauncher.createServerLauncher(server, in, out, false, printWriter);
-    } catch (FileNotFoundException | UnsupportedEncodingException e) {
-      LOGGER.error("Can't create LSP trace file", e);
-      if (logFile.isDirectory()) {
-        LOGGER.error("Trace log setting must lead to file, not directory! {}", logFile.getAbsolutePath());
-      }
-
-      launcher = LSPLauncher.createServerLauncher(server, in, out);
-    }
-
-    return launcher;
   }
 }
